Split architecture related functions to separated folder
