+++ plugins/arch/CMakeLists.txt
@@ -3,3 +3,5 @@ SET(SOURCE_GROUP ${SOURCE_GROUP}\\arch)
 add_subdirectory(posix)
 add_subdirectory(win32)
 add_subdirectory(arduino)
+add_subdirectory(freertos)
+add_subdirectory(vxworks)
