+++ plugins/arch/vxworks/CMakeLists.txt
@@ -1,0 +1,13 @@
+SET(SOURCE_GROUP ${SOURCE_GROUP}\\vxworks)
+
+ua_add_architecture("vxworks")
+
+if("${UA_ARCHITECTURE}" STREQUAL "vxworks")
+
+    ua_include_directories(${CMAKE_CURRENT_SOURCE_DIR})
+    ua_add_architecture_header(${CMAKE_CURRENT_SOURCE_DIR}/ua_architecture.h)
+    
+    ua_architecture_remove_definitions(-Werror -Wpedantic -Wno-static-in-inline -fPIC)
+    ua_architecture_add_definitions(-D_WRS_KERNEL)
+    
+endif()
