+++ tests/fuzz/generate_corpus.sh
